fix: split chat item delegation helpers
